remove duplicate game_end (#1239)
